rip code base
