+++ internal/email/email.go
@@ -1,7 +1,11 @@
 package email
 
 import (
+	"strings"
+
+	"github.com/go-playground/validator/v10"
 	"github.com/zvdv/ECSS-Lockers/internal/env"
+	"github.com/zvdv/ECSS-Lockers/internal/logger"
 	"gopkg.in/gomail.v2"
 )
 
@@ -11,12 +15,12 @@ var (
 )
 
 func init() {
-	HostEmail = env.MustEnv("EMAIL_HOST_ADDRESS")
+	HostEmail = env.Env("EMAIL_HOST_ADDRESS")
 
 	mailDialier = gomail.NewDialer("smtp.gmail.com",
 		587,
 		HostEmail,
-		env.MustEnv("EMAIL_HOST_PASSWORD"))
+		env.Env("EMAIL_HOST_PASSWORD"))
 }
 
 // NOTE: overides the "From" header, will set it to
@@ -26,4 +30,13 @@ func Send(messages ...*gomail.Message) error {
 		msg.SetHeader("From", HostEmail)
 	}
 	return mailDialier.DialAndSend(messages...)
+}
+
+func ValidUVicEmail(email string) bool {
+	err := validator.New().Var(email, "email")
+	if err == nil {
+		return strings.HasSuffix(email, "@uvic.ca")
+	}
+	logger.Error("Invalid email %s:\n%v", email, err)
+	return false
 }
