+++ templates/dash/locker_card.html
@@ -1,10 +1,10 @@
 {{if .LockerOK}}
 <form
   hx-put="/dash/api/locker/confirm"
-  class="max-h-[60vh] overflow-y-scroll w-full grid grid-cols-2 place-items-center gap-5"
+  class="max-h-[60vh] overflow-y-scroll w-full grid grid-cols-2 place-items-center gap-5 p-5"
 >
   {{range $locker := .Lockers}}
-  <div class="card shadow-xl border p-5">
+  <div class="card shadow border p-5">
     <div class="flex flex-col gap-3 justify-center items-center">
       <p>{{$locker.LockerID}}</p>
