+++ templates/index.html
@@ -1,43 +1,54 @@
-<main class="flex flex-col justify-center items-center max-w-screen-md mx-auto h-[100vh] gap-5">
-    <img class="w-56" src="/assets/imgs/ess_ring_orange_trans.png" alt="UVic ESS Logo" />
+<main
+  class="flex flex-col justify-center items-center max-w-screen-md mx-auto h-[100vh] gap-5"
+>
+  <img
+    class="w-56"
+    src="/assets/imgs/ess_ring_orange_trans.png"
+    alt="UVic ESS Logo"
+  />
 
-    <h1 class="text-3xl font-bold">Locker Registration</h1>
+  <h1 class="text-3xl font-bold">Locker Registration</h1>
 
-    <form 
-        hx-post="/api/login" 
-        hx-trigger="submit"
-        hx-swap="innerHTML"
-        hx-target="#output"
-        hx-disabled-elt="find button[type='submit'], find input[type='text']"
-        class="flex flex-col gap-4 justify-center items-center">
+  <form
+    hx-post="/auth/api/login"
+    hx-trigger="submit"
+    hx-swap="innerHTML"
+    hx-target="#output"
+    hx-disabled-elt="find button[type='submit'], find input[type='text']"
+    class="flex flex-col gap-4 justify-center items-center"
+  >
+    <div>
+      <label class="input input-bordered flex items-center gap-2">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          viewBox="0 0 16 16"
+          fill="currentColor"
+          class="h-4 w-4 opacity-70"
+        >
+          <path
+            d="M2.5 3A1.5 1.5 0 0 0 1 4.5v.793c.026.009.051.02.076.032L7.674 8.51c.206.1.446.1.652 0l6.598-3.185A.755.755 0 0 1 15 5.293V4.5A1.5 1.5 0 0 0 13.5 3h-11Z"
+          />
+          <path
+            d="M15 6.954 8.978 9.86a2.25 2.25 0 0 1-1.956 0L1 6.954V11.5A1.5 1.5 0 0 0 2.5 13h11a1.5 1.5 0 0 0 1.5-1.5V6.954Z"
+          />
+        </svg>
 
-        <div>
+        <span>
+          <input
+            name="email"
+            type="text"
+            class="grow"
+            placeholder="netlinkid"
+          />@uvic.ca
+        </span>
+      </label>
+    </div>
 
-            <label class="input input-bordered flex items-center gap-2">
-
-                <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    viewBox="0 0 16 16"
-                    fill="currentColor"
-                    class="h-4 w-4 opacity-70">
-                    <path
-                        d="M2.5 3A1.5 1.5 0 0 0 1 4.5v.793c.026.009.051.02.076.032L7.674 8.51c.206.1.446.1.652 0l6.598-3.185A.755.755 0 0 1 15 5.293V4.5A1.5 1.5 0 0 0 13.5 3h-11Z" />
-                    <path
-                        d="M15 6.954 8.978 9.86a2.25 2.25 0 0 1-1.956 0L1 6.954V11.5A1.5 1.5 0 0 0 2.5 13h11a1.5 1.5 0 0 0 1.5-1.5V6.954Z" />
-                </svg>
-
-                <span>
-                    <input name="email" type="text" class="grow" placeholder="netlinkid" />@uvic.ca
-                </span>
-
-            </label>
-
-        </div>
-
-        <div id="output" class="w-full flex flex-col gap-3 justify-center items-center">
-            <button type="submit" 
-                class="btn btn-primary btn-block">Login</button>
-        </div>
-
-    </form>
+    <div
+      id="output"
+      class="w-full flex flex-col gap-3 justify-center items-center"
+    >
+      <button type="submit" class="btn btn-primary btn-block">Login</button>
+    </div>
+  </form>
 </main>
